requirements.txt
django-suit
-r requirements-dev.txt
gunicorn
psycopg2